+++ packages/orq-rc/src/orq_ai_sdk/models/findonebykeypromptsnippetop.py
@@ -1329,7 +1329,7 @@ class FindOneByKeyPromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-04T13:37:25.030Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-05T07:31:09.342Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")